LayoutTests/fast/text/font-face-set-javascript-expected.txt

PASS new FontFaceSet() threw exception TypeError: Not enough arguments.
PASS new FontFaceSet([]).size is 0
PASS new FontFaceSet([fontFace1]).size is 1
PASS fontFaceSet.status is "loaded"
PASS item.done is false
PASS item.value is [fontFace1, fontFace1]
PASS item.done is true
PASS item.done is false
PASS item.value is fontFace1
PASS item.done is true
PASS item.done is false
PASS item.value is fontFace1
PASS item.done is true
PASS fontFaceSet.add(fontFace2) is fontFaceSet
PASS fontFaceSet.size is 2
PASS item.done is false
PASS item.value is fontFace1
PASS item.done is false
PASS item.value is fontFace2
PASS item.done is true
PASS fontFaceSet.delete(fontFace1) is true
PASS fontFaceSet.delete(fontFace3) is false
PASS fontFaceSet.size is 0
PASS fontFaceSet.values().next().done is true
PASS fontFaceSet.check('garbage') threw exception Error: SyntaxError: DOM Exception 12.
PASS fontFaceSet.check('16px garbage') is true
PASS fontFaceSet.check('16px family1') is false
PASS fontFaceSet.status is "loaded"
PASS item.code is item.SYNTAX_ERR
PASS fontFaceSet.check('16px family1') is false
PASS item is []
PASS item.code is item.NETWORK_ERR
PASS fontFaceSet.check('16px family3') is false
PASS fontFaceSet.status is "loading"
PASS item is [fontFace3]
PASS fontFaceSet.check('16px family3') is true
PASS fontFaceSet.status is "loaded"
PASS fontFaceSet.status is "loaded"
PASS item is [fontFace3]
PASS item is [fontFace3, fontFace4]
PASS item is fontFaceSet
PASS fontFaceSet.status is "loaded"
PASS fontFaceSet.status is "loading"
PASS item is fontFaceSet
PASS successfullyParsed is true

TEST COMPLETE
